Move line-related functionality into lines.py, brackets into brackets.py
